+++ src/main/java/com/jiangzhen/vo/input/PersonalInput.java
@@ -13,7 +13,7 @@ public class PersonalInput {
     @NotBlank(message = "姓名必填")
     private String personalName;
     @NotNull(message = "性别必填")
-    private Integer gender;
+    private Long gender;
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     @JsonFormat(pattern = "yyyy-MM-dd", timezone = "GMT+8")
     private Date birthday;
@@ -23,9 +23,9 @@ public class PersonalInput {
     private String education;
     private String school;
     private String imgUrl;
-    private Integer workStatus;
-    private Integer departmentId;
+    private Long workStatus;
+    private Long departmentId;
     private String departmentName;
-    private Integer positionId;
+    private Long positionId;
     private Date beginDate;
 }
